OD20DEVAT3-29, OD20DEVAT3-51, OD20DEVAT3-56, OD20DEVAT3-72, OD20DEVAT3-67, OD20DEVAT3-69 Pagination in logs. Adding new servers in ServerController. Changes LogDAO for pagination. Fixes in views. Reformat code in front-end folder.

+++ src/main/resources/frontend/angular-app/src/app/entity/Directory.ts
@@ -1,7 +1,5 @@
-import {BigInteger} from "@angular/compiler/src/i18n/big_integer";
-
 export class Directory {
-  objectId: BigInteger;
+  objectId: bigint;
   path: string;
   enabled: boolean;
   lastExistenceCheck: Date;
